feat: Add event listener to login form
